Create activation grid generator
After passing the input image through the backbone network, we have a convolutional feature map of shape  (D,7,7)  which we interpret as a  7 x 7  grid of  D− dimensional features. At each point in this grid, we predict a set of  A  bounding boxes. The total number of bounding boxes we predict for a single image are thus  A x 7 x 7 .

Centered at each position of the  7 x 7  activation grid, we predict  A  bounding boxes, where  A=2  in our case. In order to place bounding boxes centered at each position of the  7 x 7  grid of backbone features, we need to know the spatial position of the center of each cell in the grid of features.

This function will compute these center coordinates for us.