Create positive and negative bounding boxes
During training, after we calculate the bbox_xywh (offset) values for all bounding boxes, we need to match the ground-truth boxes against the predicted bounding boxes to determine the classification labels for the bounding boxes -- which boxes should be classified as containing an object and which should be classified as background? Based on this, we will caluclate the 'expected' or 'Ground Truth' targets for offsets and class, which will be used by you to caluclate the loss. We have written this part for you